pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.4",
    "earthengine-api>=1.6.9",
    "fastapi>=0.117.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "folium>=0.20.0",
    "geopandas>=1.1.1",
    "geopy>=2.4.1",
    "google-api-python-client>=2.183.0",
    "google-cloud-aiplatform>=1.116.0",
    "google-cloud-bigquery>=3.38.0",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
    "guardrails-ai>=0.6.7",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "nest-asyncio>=1.6.0",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "openai-agents>=0.3.2",
    "openai[experimental]>=1.109.1",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "uvicorn>=0.37.0",
]
